Add update User

+++ DreamStore.Core/Services/ProductService.cs
@@ -8,12 +8,7 @@ using DreamStore.Core.Specification;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
 using SixLabors.ImageSharp;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using static Org.BouncyCastle.Math.EC.ECCurve;
+
 
 namespace DreamStore.Core.Services
 {
@@ -31,7 +26,7 @@ namespace DreamStore.Core.Services
             _logger = logger;
             _config = config;
         }
-        public async Task<ServiceResponse> Create(CreateProductDto model)
+        public async Task<ServiceResponse> Create(CreateUpdateProductDto model)
         {
             var mappedProduct = _mapper.Map<AppProduct>(model);
             if (model.Photo != null)
@@ -90,6 +85,22 @@ namespace DreamStore.Core.Services
 
         public async Task<ServiceResponse> DeleteById(int Id)
         {
+            AppProduct? product = await _productRepo.GetByID(Id);
+            if (product is null)
+            {
+                return new ServiceResponse()
+                {
+                    Success = false,
+                    Message = "product not found some problem with Id",
+                };
+            }
+
+            if (product.ImageName != "default.png")
+            {
+                var productImageFolder = _config.GetValue<string>("ImageSettings:ProductImage");
+                string image = Path.Combine(Directory.GetCurrentDirectory(), productImageFolder!, product.ImageName);
+                File.Delete(image);
+            }
             try
             {
                 await _productRepo.Delete(Id);
@@ -138,9 +149,72 @@ namespace DreamStore.Core.Services
             return _mapper.Map<ProductDto>(await _productRepo.GetByID(Id));
         }
 
-        public Task<ServiceResponse> Update()
+        public async Task<ServiceResponse> Update(CreateUpdateProductDto model)
         {
-            throw new NotImplementedException();
+            AppProduct? oldProduct = await _productRepo.GetByID(model.Id);
+            if (oldProduct == null)
+            {
+                return new ServiceResponse 
+                {
+                    Success = false,
+                    Message = "Product not found"
+                };
+            }
+          
+            if (model.Photo != null)
+            {
+                try
+                {
+                    using var image = await Image.LoadAsync(model.Photo.OpenReadStream());
+                }
+                catch (UnknownImageFormatException)
+                {
+                    return new ServiceResponse
+                    {
+                        Success = false,
+                        Message = "Bad photo format"
+                    };
+                }
+                var ProductImageFolder = _config.GetValue<string>("ImageSettings:ProductImage");
+
+                if (oldProduct.ImageName != "default.png" && ProductImageFolder != null)
+                {
+                    string oldImage = Path.Combine(Directory.GetCurrentDirectory(), ProductImageFolder, oldProduct.ImageName);
+                    File.Delete(oldImage);
+                }
+
+                string fileName = Guid.NewGuid().ToString() + Path.GetExtension(model.Photo.FileName);
+                string upload = Path.Combine(Directory.GetCurrentDirectory(), ProductImageFolder!);
+
+
+                using (var fileStream = new FileStream(Path.Combine(upload, fileName), FileMode.Create))
+                {
+                    model.Photo.CopyTo(fileStream);
+                }
+
+                oldProduct = _mapper.Map<AppProduct>(model);
+                oldProduct.ImageName = fileName;
+            }
+       
+            try
+            {
+                await _productRepo.Update(oldProduct);
+                await _productRepo.Save();
+                return new ServiceResponse
+                {
+                    Success = true,
+                    Message = "Product Updated"
+                };
+            }
+            catch (Exception ex)
+            {
+                _logger.LogError(ex, "Failed to Update  Product with name {name}", model.Name);
+                return new ServiceResponse
+                {
+                    Success = false,
+                    Message = "Some problem with Update Product"
+                };
+            }
         }
     }
 }
